requirements.txt
pydantic~=2.11.7
stripe~=12.4.0
fastapi~=0.116.1
requests~=2.32.4
pillow~=11.3.0
aiogram~=3.21.0
uvicorn~=0.35.0
SQLAlchemy~=2.0.41
python-dotenv~=1.1.1